Factored SWT UI to separate class.
The implementation of the Apple //e in eric.Apple2e is now "headless".
It simulates the Apple //e hardware without any GUI.  The GUI has been
factored to a separate class, SWTUI, which creates the Apple2e virtual
machine, and periodically updates the display.

Because of the course refactoring, many private variables in Apple2e
(like the soft switch settings and the RAM) are now public variables,
rather than privates, because they are used in the UI to render the
display.

The rationale behind this refactoring is to make it possible to easily
replace the SWT-based UI with a different UI, perhaps one based on AWT,
or, more likely, as a proof of concept for a SDL-based user interface.