Migrate to gradle
The old ant-based build system was a liability, and fetching the
SWT depdency by hand was necessary years ago, but it isn't anymore.
Furthermore, the old build was desirable when using the Avian VM, which
is long dead.

This makes the project more modern and requires less boilerplate.

I was able to successfully test this with a win32 JVM. It hopefully
will work with minimal modifications on other platforms.