Add regression test suite
This adds a junit suite to exercise all the instructions in the M6502
processor. The suite comes from
  https://github.com/Klaus2m5/6502_65C02_functional_tests

The JUnit test file reads the memory image and then starts running the
test (at location 0x400). The test runs linearly, and if successful, it
branches back to itself. (It does this on failure too, but at different
locattions).

The test suite can be run with

  ./gradlew test

Note that these tests don't pass just yet because of some bugs in the
CPU emulation.