Remove pcbuf support
pcbuf was short for "program counter buffer". When debugging the
emulator, it was useful to get a trace of the most recent instructions
that had been run. This was used in crash reports.

While it may be useful to add debugging support back in at some point,
the current mechanism was no longer used.